changed type of maxNr from int to long

+++ src/main/java/nl/clockwork/ebms/dao/AbstractEbMSDAO.java
@@ -1080,10 +1080,10 @@ public abstract class AbstractEbMSDAO implements EbMSDAO
 		}
 	}
 
-	public abstract String getMessageIdsQuery(String messageContextFilter, EbMSMessageStatus status, int maxNr);
+	public abstract String getMessageIdsQuery(String messageContextFilter, EbMSMessageStatus status, long maxNr);
 
 	@Override
-	public List<String> getMessageIds(EbMSMessageContext messageContext, EbMSMessageStatus status, int maxNr) throws DAOException
+	public List<String> getMessageIds(EbMSMessageContext messageContext, EbMSMessageStatus status, long maxNr) throws DAOException
 	{
 		Connection c = null;
 		PreparedStatement ps = null;
